requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
openai>=1.3.7
pydantic>=2.9.0
jinja2>=3.1.2
python-multipart>=0.0.6
aiofiles>=23.2.1
requests>=2.32.0